pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "breadbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "colorama>=0.4.6",
    "discord-py>=2.5.2",
    "python-dotenv>=1.1.1",
    "sqlmodel>=0.0.24",
]

[project.scripts]
start = "bot.main:main"

[tool.setuptools.packages.find]
where = ["src"]
